create a component to analyze sprite shapes
