+++ Maze_Shooter/Assets/Prefabs/Tools/ground shape.prefab
@@ -11,6 +11,7 @@ GameObject:
   - component: {fileID: 2223938536542909184}
   - component: {fileID: 2223938536542909187}
   - component: {fileID: 2223938536542909186}
+  - component: {fileID: 5795025363136612377}
   - component: {fileID: 2223938536542909185}
   m_Layer: 0
   m_Name: ground shape
@@ -611,6 +612,22 @@ MonoBehaviour:
   m_EnableTangents: 0
   m_GeometryCached: 0
   m_UTess2D: 0
+--- !u!114 &5795025363136612377
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 2223938536542909191}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 49216f781df9ccb478cf14f7f5998c9b, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  spriteShapeController: {fileID: 2223938536542909186}
+  curveSegments: 5
+  simplify: 0.005
+  heightFactor: 2
 --- !u!114 &2223938536542909185
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -629,106 +646,9 @@ MonoBehaviour:
   voxelWidth: 5
   voxelHeight: 0.5
   height: 5
-  curveSegments: 6
-  simplify: 0.005
   thickness: -2
-  spriteShapeController: {fileID: 2223938536542909186}
-  points:
-  - {x: 0.8288894, y: 79.93255, z: 0.00000004304006}
-  - {x: 1.1117618, y: 83.20341, z: 0.000000022788406}
-  - {x: 1.1117618, y: 83.20341, z: 0.000000022788406}
-  - {x: -5.9207606, y: 83.13357, z: 0.000000013704976}
-  - {x: -5.9207606, y: 83.13357, z: 0.000000013704976}
-  - {x: -8.933509, y: 83.103645, z: 0.0000000098042845}
-  - {x: -8.933509, y: 83.103645, z: 0.0000000098042845}
-  - {x: -14.81331, y: 83.04526, z: 0.0000000022457656}
-  - {x: -14.81331, y: 83.04526, z: 0.0000000022457656}
-  - {x: -14.966232, y: 79.46283, z: 0.00000006263189}
-  - {x: -14.966232, y: 79.46283, z: 0.00000006263189}
-  - {x: -45.00836, y: 78.071014, z: 0.0000000026592488}
-  - {x: -45.00836, y: 78.071014, z: 0.0000000026592488}
-  - {x: -44.087772, y: 85.92139, z: 0.0000000023985114}
-  - {x: -44.087772, y: 85.92139, z: 0.0000000023985114}
-  - {x: -32.076702, y: 86.88643, z: -0.000000012591267}
-  - {x: -32.076702, y: 86.88643, z: -0.000000012591267}
-  - {x: -28.767286, y: 87.11252, z: 0.00000019231585}
-  - {x: -28.767286, y: 87.11252, z: 0.00000019231585}
-  - {x: -26.347786, y: 87.29812, z: 0.00000014430013}
-  - {x: -26.347786, y: 87.29812, z: 0.00000014430013}
-  - {x: -19.184464, y: 87.8476, z: 0.0000000021157522}
-  - {x: -19.184464, y: 87.8476, z: 0.0000000021157522}
-  - {x: -18.949202, y: 91.18684, z: -0.000000075787696}
-  - {x: -18.949202, y: 91.18684, z: -0.000000075787696}
-  - {x: -15.416726, y: 91.15934, z: -0.00000014260603}
-  - {x: -15.416726, y: 91.15934, z: -0.00000014260603}
-  - {x: -8.861949, y: 91.108315, z: -0.00000026662542}
-  - {x: -8.861949, y: 91.108315, z: -0.00000026662542}
-  - {x: -5.2476983, y: 90.9605, z: 0.0000000065762}
-  - {x: -5.2476983, y: 90.9605, z: 0.0000000065762}
-  - {x: 1.4742022, y: 90.849464, z: 0.000000046968875}
-  - {x: 1.4742022, y: 90.849464, z: 0.000000046968875}
-  - {x: 1.7246256, y: 94.10663, z: 0.000000037368473}
-  - {x: 1.7246256, y: 94.10663, z: 0.000000037368473}
-  - {x: 5.21471, y: 94.16931, z: -0.000000031603857}
-  - {x: 5.21471, y: 94.16931, z: -0.000000031603857}
-  - {x: 6.9429927, y: 104.91432, z: -0.000000013628395}
-  - {x: 6.9429927, y: 104.91432, z: -0.000000013628395}
-  - {x: 8.085026, y: 111.02367, z: -0.000000078097}
-  - {x: 8.085026, y: 111.02367, z: -0.000000078097}
-  - {x: 8.468028, y: 114.39577, z: 0.0000000022600277}
-  - {x: 8.468028, y: 114.39577, z: 0.0000000022600277}
-  - {x: 8.613063, y: 135.12921, z: 0.0000000038198777}
-  - {x: 8.613063, y: 135.12921, z: 0.0000000038198777}
-  - {x: 5.3174305, y: 135.0851, z: 0.000000055266874}
-  - {x: 5.3174305, y: 135.0851, z: 0.000000055266874}
-  - {x: 5.1155148, y: 163.79288, z: 0.000000004785761}
-  - {x: 5.1155148, y: 163.79288, z: 0.000000004785761}
-  - {x: 7.851017, y: 163.87349, z: -0.000000018145384}
-  - {x: 7.851017, y: 163.87349, z: -0.000000018145384}
-  - {x: 7.886626, y: 176.53738, z: 0.000000031071053}
-  - {x: 7.886626, y: 176.53738, z: 0.000000031071053}
-  - {x: 7.8708463, y: 186.96513, z: 0.00000003785852}
-  - {x: 7.8708463, y: 186.96513, z: 0.00000003785852}
-  - {x: 13.725772, y: 174.0351, z: 0.00000018849128}
-  - {x: 13.725772, y: 174.0351, z: 0.00000018849128}
-  - {x: 13.218925, y: 165.0647, z: 0.0000000065943286}
-  - {x: 13.218925, y: 165.0647, z: 0.0000000065943286}
-  - {x: 13.211204, y: 162.45847, z: -0.00000006839885}
-  - {x: 13.211204, y: 162.45847, z: -0.00000006839885}
-  - {x: 14.145358, y: 153.16013, z: 0.0000000015962196}
-  - {x: 14.145358, y: 153.16013, z: 0.0000000015962196}
-  - {x: 14.392606, y: 149.98279, z: 3.170339e-10}
-  - {x: 14.392606, y: 149.98279, z: 3.170339e-10}
-  - {x: 15.027702, y: 141.82178, z: -0.0000000030939131}
-  - {x: 15.027702, y: 141.82178, z: -0.0000000030939131}
-  - {x: 12.228407, y: 141.71083, z: 0.000000005621132}
-  - {x: 12.228407, y: 141.71083, z: 0.000000005621132}
-  - {x: 12.205962, y: 139.04506, z: 0.000000005250513}
-  - {x: 12.205962, y: 139.04506, z: 0.000000005250513}
-  - {x: 16.05217, y: 139.12282, z: 0.0000000042429824}
-  - {x: 16.05217, y: 139.12282, z: 0.0000000042429824}
-  - {x: 16.110788, y: 131.94128, z: 0.00000003230555}
-  - {x: 16.110788, y: 131.94128, z: 0.00000003230555}
-  - {x: 18.73584, y: 131.72238, z: -0.00000020903303}
-  - {x: 18.73584, y: 131.72238, z: -0.00000020903303}
-  - {x: 18.337242, y: 113.83839, z: -0.000000051703623}
-  - {x: 18.337242, y: 113.83839, z: -0.000000051703623}
-  - {x: 15.076814, y: 113.78335, z: 0.0000000023512712}
-  - {x: 15.076814, y: 113.78335, z: 0.0000000023512712}
-  - {x: 12.35551, y: 93.852516, z: -0.00000003915584}
-  - {x: 12.35551, y: 93.852516, z: -0.00000003915584}
-  - {x: 15.599462, y: 93.66861, z: -0.00000007948643}
-  - {x: 15.599462, y: 93.66861, z: -0.00000007948643}
-  - {x: 15.588072, y: 90.390236, z: -0.000000104314886}
-  - {x: 15.588072, y: 90.390236, z: -0.000000104314886}
-  - {x: 35.046288, y: 90.35493, z: -0.0000005506364}
-  - {x: 35.046288, y: 90.35493, z: -0.0000005506364}
-  - {x: 19.32259, y: 84.25865, z: -0.00000011287274}
-  - {x: 19.32259, y: 84.25865, z: -0.00000011287274}
-  - {x: 13.595291, y: 84.197815, z: -0.000000022079469}
-  - {x: 13.595291, y: 84.197815, z: -0.000000022079469}
-  - {x: 13.1366005, y: 79.73491, z: 0.0000000072342274}
-  - {x: 13.1366005, y: 79.73491, z: 0.0000000072342274}
-  - {x: 0.8288894, y: 79.93255, z: 0.00000004304006}
+  thicknessFromShape: 1
+  offset: 0.5
+  overlap: 0.15
   walls: []
   voxels: []
