+++ Maze_Shooter/Assets/Scripts/sprite shape extensions.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: fcb0b1f4fac79dc448aedfa8a02ebeaa
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
